requirements.txt
google-cloud-speech>=2.32.0
google-auth>=2.40.2
protobuf>=5.29.4
grpcio>=1.71.0
sounddevice>=0.4.6
numpy>=1.26.0
deepseek>=1.0.0
python-dotenv>=1.0.0
